Ensure add_attributes handles merging existing attributes.
Emulates what Drupal supports, where add_attributes will merge with any attributes that have already been set rather than overwriting them.

Note that the attributes set in the component file are treated as the "new" attributes and anything passed into the template in the attributes variable is the "old" or "existing" attributes.